+++ src/llm/backends/anthropic.py
@@ -32,11 +32,9 @@ class AnthropicBackend:
         thinking_budget_tokens: int | None = None,
         thinking_effort: str | None = None,
         max_output_tokens: int | None = None,
-        api_key: str | None = None,
-        api_base: str | None = None,
         extra_params: dict[str, Any] | None = None,
     ) -> CompletionResult:
-        del max_output_tokens, api_key, api_base
+        del max_output_tokens
         if thinking_effort is not None:
             raise ValueError(
                 "Anthropic backend does not support thinking_effort; use thinking_budget_tokens instead"
@@ -122,12 +120,10 @@ class AnthropicBackend:
         thinking_budget_tokens: int | None = None,
         thinking_effort: str | None = None,
         max_output_tokens: int | None = None,
-        api_key: str | None = None,
-        api_base: str | None = None,
         extra_params: dict[str, Any] | None = None,
     ) -> AsyncIterator[StreamChunk]:
         is_json_mode = self._json_mode(extra_params)
-        del max_output_tokens, api_key, api_base
+        del max_output_tokens
         if thinking_effort is not None:
             raise ValueError(
                 "Anthropic backend does not support thinking_effort; use thinking_budget_tokens instead"
