+++ tests/conftest.py
@@ -714,10 +714,10 @@ def mock_honcho_llm_call(request: pytest.FixtureRequest):
     # Patch the honcho_llm_call decorator to prevent actual LLM calls at module level
     original_decorator = None
     try:
-        import src.utils.clients
+        import src.llm
 
-        original_decorator = src.utils.clients.honcho_llm_call
-        src.utils.clients.honcho_llm_call = lambda *args, **kwargs: lambda func: func  # pyright: ignore[reportUnknownLambdaType]
+        original_decorator = src.llm.honcho_llm_call
+        src.llm.honcho_llm_call = lambda *args, **kwargs: lambda func: func  # pyright: ignore[reportUnknownLambdaType]
     except ImportError:
         pass
 
@@ -751,15 +751,15 @@ def mock_honcho_llm_call(request: pytest.FixtureRequest):
 
         return mock_llm_decorator
 
-    with patch("src.utils.clients.honcho_llm_call", side_effect=decorator_factory):
+    with patch("src.llm.honcho_llm_call", side_effect=decorator_factory):
         yield decorator_factory
 
     # Restore the original decorator
     if original_decorator:
         try:
-            import src.utils.clients
+            import src.llm
 
-            src.utils.clients.honcho_llm_call = original_decorator
+            src.llm.honcho_llm_call = original_decorator
         except ImportError:
             pass
